requirements.txt
zstandard==0.23.0
ninja==1.11.1.1
wandb==0.18.5
numpy==1.26.3
tiktoken==0.7.0
blobfile==3.0.0
lm-eval==0.4.4
matplotlib==3.9.2
openai==1.54.3
orjson==3.10.11
sentencepiece==0.2.0
inquirerpy==0.3.4
anthropic==0.39.0
math-verify[antlr4_13_2]==0.5.2
sacrebleu==2.4.3
datasets==2.21.0
datatrove==0.3.0
rouge-score==0.1.2
msgspec==0.18.6
tenacity==9.0.0
seaborn==0.13.2
tqdm==4.66.6
trl==0.13.0
python-dotenv==1.0.1
lm-eval==0.4.4
scipy==1.14.1
omegaconf==2.4.0.dev1
fsspec==2024.2.0
viztracer==0.17.0
cmake==3.31.0.1
websockets==14.1
uvloop==0.21.0
transformers==4.46.2
peft==0.13.2
accelerate==1.2.1
einops==0.8.0
huggingface-hub==0.26.2
safetensors==0.4.5
tokenizers==0.20.3
lm-format-enforcer==0.10.6
pynvml